+++ lib/iex/api.rb
@@ -17,8 +17,16 @@ module IEX
   end
 end
 
-require_relative 'api/quote'
-require_relative 'api/company'
-require_relative 'api/price'
-require_relative 'api/news'
 require_relative 'api/chart'
+require_relative 'api/company'
+require_relative 'api/dividends'
+require_relative 'api/earnings'
+require_relative 'api/largest_trades'
+require_relative 'api/logo'
+require_relative 'api/key_stats'
+require_relative 'api/news'
+require_relative 'api/ohlc'
+require_relative 'api/price'
+require_relative 'api/quote'
+require_relative 'api/sectors'
+require_relative 'api/crypto'