+++ lib/iex/resources/resource.rb
@@ -1,4 +1,5 @@
 module IEX
   class Resource < Hashie::Trash
+    include Hashie::Extensions::IgnoreUndeclared
   end
 end